Preserve data in json file
